# Faster collapse