--- a/openquake/baselib/performance.py
+++ b/openquake/baselib/performance.py
@@ -192,7 +192,7 @@ class Monitor(object):
         self.inner_loop = inner_loop
         self.h5 = h5
         self.version = version
-        self.mem = 0
+        self._mem = 0
         self.duration = 0
         self._start_time = self._stop_time = time.time()
         self.children = []
@@ -200,6 +200,11 @@ class Monitor(object):
         self.address = None
         self.username = getpass.getuser()
         self.task_no = -1  # overridden in parallel
+
+    @property
+    def mem(self):
+        """Mean memory allocation"""
+        return self._mem / self.counts
 
     @property
     def dt(self):
@@ -247,7 +252,7 @@ class Monitor(object):
         self.exc = exc
         if self.measuremem:
             self.stop_mem = self.measure_mem()
-            self.mem += self.stop_mem - self.start_mem
+            self._mem += self.stop_mem - self.start_mem
         self._stop_time = time.time()
         self.duration += self._stop_time - self._start_time
         self.counts += 1
@@ -274,7 +279,7 @@ class Monitor(object):
         Reset duration, mem, counts
         """
         self.duration = 0
-        self.mem = 0
+        self._mem = 0
         self.counts = 0
 
     def flush(self, h5):
@@ -349,7 +354,7 @@ class Monitor(object):
         """
         while True:
             try:
-                self.mem = 0
+                self._mem = 0
                 with self:
                     obj = next(genobj)
             except StopIteration: