Here is an example for a reduced AUS computation on the spot machine (21% speedup on collapsing):
```
# before
| calc_847, maxmem=212.4 GB  | time_sec | memory_mb | counts     |
|----------------------------+----------+-----------+------------|
| total classical            | 92_191   | 756.8     | 292        |
| collapsing contexts        | 32_771   | 9_704     | 3_389      |
| get_poes                   | 21_183   | 0.0       | 22_005_788 |
| composing pnes             | 21_111   | 0.0       | 22_005_788 |
| planar contexts            | 8_934    | 0.0       | 284_154    |
| computing mean_std         | 5_733    | 0.0       | 219_594    |
| ClassicalCalculator.run    | 1_489    | 3_129     | 1          |

# after
| calc_846, maxmem=211.7 GB  | time_sec | memory_mb | counts     |
|----------------------------+----------+-----------+------------|
| total classical            | 86_722   | 878.7     | 292        |
| collapsing contexts        | 26_973   | 10_932    | 3_389      |
| get_poes                   | 21_252   | 0.0       | 22_005_788 |
| composing pnes             | 21_248   | 0.0       | 22_005_788 |
| planar contexts            | 8_965    | 0.0       | 284_154    |
| computing mean_std         | 5_782    | 0.0       | 219_594    |
| ClassicalCalculator.run    | 1_370    | 3_136     | 1          |
```